+++ web-wallet/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2024-02-28
+
 ### Added
 - Add Create Wallet flow tests [#1443]
 - Add visible version, commit hash and build date [#1441]
@@ -118,6 +120,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- VERSIONS -->
 [Unreleased]: https://github.com/dusk-network/rusk/tree/master/web-wallet
+[0.3.0]: https://github.com/dusk-network/rusk/tree/web-wallet-0.3.0
 [0.2.1]: https://github.com/dusk-network/rusk/tree/web-wallet-0.2.1
 [0.2.0]: https://github.com/dusk-network/rusk/tree/web-wallet-0.2.0
 [0.1.0-beta]: https://github.com/dusk-network/rusk/tree/web-wallet-0.1.0-beta